Cleanup of allResults
