+++ src/allResults.html
@@ -30,5 +30,5 @@
 
 
 </body>
-<script src="allResults.js"></script>
+<script src="allResults.ts"></script>
 </html>
